+++ media/webrtc/trunk/webrtc/common_audio/vad/vad_core.h
@@ -104,12 +104,12 @@ int WebRtcVad_set_mode_core(VadInstT* self, int mode);
 
 
 int WebRtcVad_CalcVad48khz(VadInstT* inst, const int16_t* speech_frame,
-                           int frame_length);
+                           size_t frame_length);
 int WebRtcVad_CalcVad32khz(VadInstT* inst, const int16_t* speech_frame,
-                           int frame_length);
+                           size_t frame_length);
 int WebRtcVad_CalcVad16khz(VadInstT* inst, const int16_t* speech_frame,
-                           int frame_length);
+                           size_t frame_length);
 int WebRtcVad_CalcVad8khz(VadInstT* inst, const int16_t* speech_frame,
-                          int frame_length);
+                          size_t frame_length);
 
 #endif
