+++ media/webrtc/trunk/webrtc/common_audio/vad/vad_sp.h
@@ -33,7 +33,7 @@
 void WebRtcVad_Downsampling(const int16_t* signal_in,
                             int16_t* signal_out,
                             int32_t* filter_state,
-                            int in_length);
+                            size_t in_length);
