+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/ilbc/cb_search_core.h
@@ -23,13 +23,13 @@
 
 void WebRtcIlbcfix_CbSearchCore(
     int32_t *cDot,    
-    int16_t range,    
+    size_t range,    
     int16_t stage,    
     int16_t *inverseEnergy,  
     int16_t *inverseEnergyShift, 
 
     int32_t *Crit,    
-    int16_t *bestIndex,   
+    size_t *bestIndex,   
 
 
     int32_t *bestCrit,
