+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/ilbc/decode.c
@@ -44,7 +44,7 @@ void WebRtcIlbcfix_DecodeImpl(
     int16_t mode      
 
                            ) {
-  int i;
+  size_t i;
   int16_t order_plus_one;
 
   int16_t last_bit;
@@ -103,9 +103,10 @@ void WebRtcIlbcfix_DecodeImpl(
       WebRtcIlbcfix_DecodeResidual(iLBCdec_inst, iLBCbits_inst, decresidual, syntdenum);
 
       
-      WebRtcIlbcfix_DoThePlc( PLCresidual, PLClpc, 0,
-                              decresidual, syntdenum + (LPC_FILTERORDER + 1)*(iLBCdec_inst->nsub - 1),
-                              (int16_t)(iLBCdec_inst->last_lag), iLBCdec_inst);
+      WebRtcIlbcfix_DoThePlc(
+          PLCresidual, PLClpc, 0, decresidual,
+          syntdenum + (LPC_FILTERORDER + 1) * (iLBCdec_inst->nsub - 1),
+          iLBCdec_inst->last_lag, iLBCdec_inst);
 
       
       WEBRTC_SPL_MEMCPY_W16(decresidual, PLCresidual, iLBCdec_inst->blockl);
@@ -120,8 +121,8 @@ void WebRtcIlbcfix_DecodeImpl(
 
     
 
-    WebRtcIlbcfix_DoThePlc( PLCresidual, PLClpc, 1,
-                            decresidual, syntdenum, (int16_t)(iLBCdec_inst->last_lag), iLBCdec_inst);
+    WebRtcIlbcfix_DoThePlc(PLCresidual, PLClpc, 1, decresidual, syntdenum,
+                           iLBCdec_inst->last_lag, iLBCdec_inst);
 
     WEBRTC_SPL_MEMCPY_W16(decresidual, PLCresidual, iLBCdec_inst->blockl);
 
@@ -187,18 +188,18 @@ void WebRtcIlbcfix_DecodeImpl(
     WEBRTC_SPL_MEMCPY_W16(iLBCdec_inst->syntMem, &data[iLBCdec_inst->blockl-LPC_FILTERORDER], LPC_FILTERORDER);
 
   } else { 
-    int16_t lag;
+    size_t lag;
 
     
     lag = 20;
     if (iLBCdec_inst->mode==20) {
-      lag = (int16_t)WebRtcIlbcfix_XcorrCoef(
+      lag = WebRtcIlbcfix_XcorrCoef(
           &decresidual[iLBCdec_inst->blockl-60],
           &decresidual[iLBCdec_inst->blockl-60-lag],
           60,
           80, lag, -1);
     } else {
-      lag = (int16_t)WebRtcIlbcfix_XcorrCoef(
+      lag = WebRtcIlbcfix_XcorrCoef(
           &decresidual[iLBCdec_inst->blockl-ENH_BLOCKL],
           &decresidual[iLBCdec_inst->blockl-ENH_BLOCKL-lag],
           ENH_BLOCKL,
@@ -206,7 +207,7 @@ void WebRtcIlbcfix_DecodeImpl(
     }
 
     
-    (*iLBCdec_inst).last_lag = (int)lag;
+    (*iLBCdec_inst).last_lag = lag;
 
     
     WEBRTC_SPL_MEMCPY_W16(data, decresidual, iLBCdec_inst->blockl);
