+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/ilbc/energy_inverse.h
@@ -26,7 +26,7 @@
 void WebRtcIlbcfix_EnergyInverse(
     int16_t *energy,     
 
-    int noOfEnergies);   
+    size_t noOfEnergies);   
 
 
 #endif
