+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/ilbc/init_decode.h
@@ -25,7 +25,7 @@
 
 
 
-int16_t WebRtcIlbcfix_InitDecode(  
+int WebRtcIlbcfix_InitDecode(  
     IlbcDecoder *iLBCdec_inst, 
     int16_t mode,     
     int use_enhancer
