+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/ilbc/init_encode.h
@@ -25,7 +25,7 @@
 
 
 
-int16_t WebRtcIlbcfix_InitEncode(  
+int WebRtcIlbcfix_InitEncode(  
     IlbcEncoder *iLBCenc_inst, 
     int16_t mode     
                                          );
