+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/ilbc/interpolate_samples.h
@@ -28,7 +28,7 @@
 void WebRtcIlbcfix_InterpolateSamples(
     int16_t *interpSamples, 
     int16_t *CBmem,   
-    int16_t lMem    
+    size_t lMem    
                                       );
 
 #endif
